Create new, dedicated weather scope.
