+++ views.json
@@ -30,6 +30,16 @@
     },
     {
       "elements": [
+        "Groundspeed",
+        "Altitude",
+        "Weather Scope",
+        "Compass",
+        "GPS Not Available"
+      ],
+      "name": "Weather Scope"
+    },
+    {
+      "elements": [
         "Compass",
         "ADSB Target Bugs Only",
         "ADSB Target Bugs",
